Change UID/GID of www-data on Linux systems
The document root gets mounted from the host into the container and
therefore has the same uid and gid of the user on the host. These IDs
are not available inside the container and PHP is running as www-data
and therefore not allowed to write to these directorys.

This problem has not accured before because this tool has only been used
on macOS where Docker is run inside an additional virtual machine which
does some magic to fix the permissions for mounted volumes.

The new script detects if the host is running Linux and then changes the
UID/GID of the www-data user inside the container to the values of the
original user on the host. Therefore the files under '/local' belong to
www-data and craft can write its files to this location.