Updated to Ubuntu Focal, MySQL 5.7 and new default PHP Version
This commit upgrades the MySQL server from 5.6 to 5.7. After you started
the new database version for the first time you have to  migrate the
data to the new version before you can use it. To do this, run `docker exec -it mysql
mysql_upgrade -uroot -p` and enter `root` as password. After this has
completed, run a `stopall` and `start` to fully restart the databse
after the upgrade. Everything should now work as before.

Due to PHP 7.0 beeing EOL for a long time now, we decidet to default to
7.4 if you don't specify which version you want to use.