Use the correct user for all operations in the container
Formerly everything was run as root in the container (scripts and
shell), which resulted in wrong file ownerships which broke some
functionallity. Now the scripts as well as `craft shell` will be
executed as the www-data user to mitigate this. You will also have to
run `craft selfupdate` to get the new script